Generates Reports
Finalized the generate reports function where the student report pdf is made based on the students information for that month. It shows positive and negative taps gained throughout the month alongside the proper character equation information. The pdf's are currently being output to a reports folder in the backend folder, but it will be changed to either be placed in an email directly or downloaded locally. Also may have added a functioning tap tracker.